+++ src/main/java/UnitInch.java
@@ -1,6 +1,6 @@
 public class UnitInch extends Unit {
     @Override
-    public int rateToInch() {
+    public double rateToInch() {
         return 1;
     }
 }
